refactor(tui): split diff_pane and scroll test modules out of their source files
rinkaku's own file-size check (ADR 0028) flagged diff_pane.rs (1120
lines) as warn after the header feature's additions; scroll.rs
(1032 lines) crossed the same warn threshold. Both were
production-code-small (340 and 499 lines respectively) with an
oversized in-module test block, so each follows the existing
#[path = "..."] split precedent (PR #82/#93/#95) rather than a
mechanical top/bottom cut, grouped by which function/concern each
block already pinned via its own "// ---" comments.
