+++ rinkaku-tui/src/ui/diff_pane.rs
@@ -339,782 +339,5 @@ pub(crate) fn marker_span(kind: DiffLineKind, bg: Option<Color>) -> Span<'static
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::app::{App, BlastRadiusSelection};
-    use crate::ui::draw;
-    use ratatui::Terminal;
-    use ratatui::backend::TestBackend;
-    use rinkaku_core::diff::LineRange;
-    use rinkaku_core::extract::{Classification, ExtractedSymbol, SymbolKind};
-    use rinkaku_core::graph::SymbolGraph;
-    use rinkaku_core::render::FileReport;
-
-    // --- diff_pane_header_lines (pure helper) ---
-
-    #[test]
-    fn should_join_symbol_name_and_path_on_first_header_line_when_selection_name_is_present() {
-        let stats = ChangeStats::default();
-
-        let actual = diff_pane_header_lines(Some("foo"), "src/lib.rs", &stats, 80);
-
-        assert_eq!(
-            vec![Line::styled(
-                "foo · src/lib.rs".to_string(),
-                Style::default().add_modifier(Modifier::BOLD)
-            )],
-            actual
-        );
-    }
-
-    #[test]
-    fn should_show_bare_path_on_first_header_line_when_no_selection_name() {
-        let stats = ChangeStats::default();
-
-        let actual = diff_pane_header_lines(None, "src/lib.rs", &stats, 80);
-
-        assert_eq!(
-            vec![Line::styled(
-                "src/lib.rs".to_string(),
-                Style::default().add_modifier(Modifier::BOLD)
-            )],
-            actual
-        );
-    }
-
-    #[test]
-    fn should_add_change_stats_line_when_stats_has_ranges_and_counts() {
-        let stats = ChangeStats {
-            ranges: vec![(23, 41), (57, 60)],
-            added: 18,
-            removed: 4,
-        };
-
-        let actual = diff_pane_header_lines(Some("foo"), "src/lib.rs", &stats, 80);
-
-        assert_eq!(
-            vec![
-                Line::styled(
-                    "foo · src/lib.rs".to_string(),
-                    Style::default().add_modifier(Modifier::BOLD)
-                ),
-                Line::styled(
-                    "chg: 23-41, 57-60 (+18/-4)".to_string(),
-                    Style::default().add_modifier(Modifier::DIM)
-                ),
-            ],
-            actual
-        );
-    }
-
-    #[test]
-    fn should_format_single_line_range_without_a_dash() {
-        let stats = ChangeStats {
-            ranges: vec![(5, 5)],
-            added: 1,
-            removed: 0,
-        };
-
-        let actual = diff_pane_header_lines(Some("foo"), "src/lib.rs", &stats, 80);
-
-        assert_eq!(
-            vec![
-                Line::styled(
-                    "foo · src/lib.rs".to_string(),
-                    Style::default().add_modifier(Modifier::BOLD)
-                ),
-                Line::styled(
-                    "chg: 5 (+1/-0)".to_string(),
-                    Style::default().add_modifier(Modifier::DIM)
-                ),
-            ],
-            actual
-        );
-    }
-
-    #[test]
-    fn should_omit_change_stats_line_when_stats_is_entirely_empty() {
-        let stats = ChangeStats::default();
-
-        let actual = diff_pane_header_lines(Some("foo"), "src/lib.rs", &stats, 80);
-
-        assert_eq!(1, actual.len());
-    }
-
-    #[test]
-    fn should_show_counts_without_ranges_when_ranges_is_empty_but_counts_are_nonzero() {
-        // A pure-deletion selection: `ChangeStats::ranges` excludes the
-        // zero-width deletion range (`change_stats`'s own doc comment), but
-        // the removed count is still real and worth reporting.
-        let stats = ChangeStats {
-            ranges: vec![],
-            added: 0,
-            removed: 2,
-        };
-
-        let actual = diff_pane_header_lines(Some("foo"), "src/lib.rs", &stats, 80);
-
-        assert_eq!(
-            vec![
-                Line::styled(
-                    "foo · src/lib.rs".to_string(),
-                    Style::default().add_modifier(Modifier::BOLD)
-                ),
-                Line::styled(
-                    "chg: (+0/-2)".to_string(),
-                    Style::default().add_modifier(Modifier::DIM)
-                ),
-            ],
-            actual
-        );
-    }
-
-    #[test]
-    fn should_truncate_first_header_line_keeping_the_tail_when_it_overflows_width() {
-        let stats = ChangeStats::default();
-
-        let actual = diff_pane_header_lines(
-            Some("very_long_symbol_name_here"),
-            "src/very/deeply/nested/module/lib.rs",
-            &stats,
-            20,
-        );
-
-        assert_eq!(
-            vec![Line::styled(
-                "…ested/module/lib.rs".to_string(),
-                Style::default().add_modifier(Modifier::BOLD)
-            )],
-            actual
-        );
-    }
-
-    fn symbol(id: &str, name: &str) -> ExtractedSymbol {
-        ExtractedSymbol {
-            id: id.to_string(),
-            name: name.to_string(),
-            kind: SymbolKind::Function,
-            signature: format!("fn {name}()"),
-            range: LineRange { start: 1, end: 1 },
-            container: None,
-            referenced_names: vec![],
-            dependencies: vec![],
-            omitted_dependency_matches: 0,
-            is_test: false,
-            classification: None,
-            previous_signature: None,
-        }
-    }
-
-    fn report_with_one_symbol() -> Report {
-        Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::foo", "foo")],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            fan_ins: vec![],
-            file_size_warnings: vec![],
-            file_size_bands: vec![],
-            removed: vec![],
-        }
-    }
-
-    fn diff_content_for(
-        report: &Report,
-        diff_files: &[crate::diff_view::FileHunks],
-        app: &App,
-    ) -> crate::diff_shape::DiffPaneContent {
-        crate::diff_shape::build_diff_pane_content(
-            report,
-            diff_files,
-            app.selected_diff_target(report).as_ref(),
-        )
-    }
-
-    fn buffer_text(terminal: &Terminal<TestBackend>) -> String {
-        let buffer = terminal.backend().buffer();
-        let area = buffer.area;
-        (0..area.height)
-            .map(|y| {
-                (0..area.width)
-                    .map(|x| buffer[(x, y)].symbol().to_string())
-                    .collect::<String>()
-            })
-            .collect::<Vec<_>>()
-            .join("\n")
-    }
-
-    fn find_cell_style(terminal: &Terminal<TestBackend>, line_needle: &str, token: &str) -> Style {
-        let buffer = terminal.backend().buffer();
-        let area = buffer.area;
-        for y in 0..area.height {
-            let row: String = (0..area.width)
-                .map(|x| buffer[(x, y)].symbol().to_string())
-                .collect();
-            let Some(needle_byte_offset) = row.find(line_needle) else {
-                continue;
-            };
-            let Some(token_byte_offset) = row[needle_byte_offset..].find(token) else {
-                continue;
-            };
-            let byte_offset = needle_byte_offset + token_byte_offset;
-            let column = row[..byte_offset].chars().count() as u16;
-            return buffer[(column, y)].style();
-        }
-        panic!("expected to find {token:?} within a row containing {line_needle:?}");
-    }
-
-    #[test]
-    fn should_draw_diff_pane_with_hunk_lines_when_toggled_on_a_symbol_row() {
-        let report = report_with_one_symbol();
-        // ADR 0020 defaults the right pane to Diff already, so no
-        // `ToggleDiff` press is needed to reach it here.
-        let app = App::new(&report).handle_key(crate::app::InputKey::Down);
-        let diff_text = "\
-diff --git a/lib.rs b/lib.rs
-index e69de29..4b825dc 100644
---- a/lib.rs
-+++ b/lib.rs
-@@ -1,1 +1,2 @@
- fn a() {}
-+fn foo() {}
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        let text = buffer_text(&terminal);
-        assert!(text.contains("Diff"));
-        assert!(text.contains("+fn foo() {}"));
-    }
-
-    // Dynamic-verification note (see CLAUDE.md's reviewing-changes
-    // section): this pins that a skipped file's diff pane still resolves
-    // real hunks from the raw diff text — `App::selected_diff_target`
-    // scopes a `NodeKind::File` row to `DiffTarget::File { path }`
-    // regardless of `skip_reason` (see the `app.rs` unit test
-    // `should_return_file_diff_target_when_cursor_is_on_a_skipped_file_row`),
-    // and `draw_diff_pane` looks hunks up by that path alone — so a
-    // skipped file (which has no `FileReport`/symbols to key off of) must
-    // not silently fall back to the "no diff hunks found" placeholder just
-    // because rinkaku didn't extract symbols from it.
-    #[test]
-    fn should_draw_diff_pane_with_hunk_lines_when_toggled_on_a_skipped_file_row() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            skipped: vec![rinkaku_core::render::SkippedFile {
-                path: "assets/logo.png".to_string(),
-                reason: rinkaku_core::render::SkipReason::Binary,
-            }],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            fan_ins: vec![],
-            file_size_warnings: vec![],
-            file_size_bands: vec![],
-            removed: vec![],
-            files: vec![],
-        };
-        // Row 0 is the collapsing "assets" dir; row 1 is the skipped file.
-        // ADR 0020 defaults the right pane to Diff already, so no
-        // `ToggleDiff` press is needed to reach it here.
-        let app = App::new(&report).handle_key(crate::app::InputKey::Down);
-        let diff_text = "\
-diff --git a/assets/logo.png b/assets/logo.png
-index e69de29..4b825dc 100644
-Binary files a/assets/logo.png and b/assets/logo.png differ
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        let text = buffer_text(&terminal);
-        // A binary file has no hunks at all in the diff text itself (git
-        // reports "Binary files ... differ" instead of `@@` hunks), so the
-        // correct, honest behavior is the pane's own "no diff hunks found"
-        // placeholder — this test's real assertion is that it names the
-        // right path, not a stale/mismatched one, confirming the lookup
-        // reached this row's `path` at all. Checked as two substrings
-        // rather than the whole phrase since it wraps across rendered
-        // lines at this terminal's pane width.
-        assert!(text.contains("no diff hunks found for"));
-        assert!(text.contains("assets/logo.png"));
-    }
-
-    /// Sibling of the binary-skip test above, using an unsupported-language
-    /// skip (a real text file with real hunks in the raw diff) to confirm
-    /// the diff pane actually renders content — not just the placeholder —
-    /// for a skipped-but-textual file.
-    #[test]
-    fn should_draw_diff_pane_with_hunk_lines_for_an_unsupported_language_skipped_file() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            skipped: vec![rinkaku_core::render::SkippedFile {
-                path: "vendor/lib.zig".to_string(),
-                reason: rinkaku_core::render::SkipReason::UnsupportedLanguage,
-            }],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            fan_ins: vec![],
-            file_size_warnings: vec![],
-            file_size_bands: vec![],
-            removed: vec![],
-            files: vec![],
-        };
-        // Row 0 is the collapsing "vendor" dir; row 1 is the skipped file.
-        // ADR 0020 defaults the right pane to Diff already, so no
-        // `ToggleDiff` press is needed to reach it here.
-        let app = App::new(&report).handle_key(crate::app::InputKey::Down);
-        let diff_text = "\
-diff --git a/vendor/lib.zig b/vendor/lib.zig
-index e69de29..4b825dc 100644
---- a/vendor/lib.zig
-+++ b/vendor/lib.zig
-@@ -1,1 +1,2 @@
- const a = 1;
-+const b = 2;
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        let text = buffer_text(&terminal);
-        assert!(text.contains("Diff"));
-        assert!(text.contains("+const b = 2;"));
-    }
-
-    #[test]
-    fn should_draw_per_symbol_section_headers_when_diff_pane_shows_a_file_selection() {
-        // Cursor stays on row 0, the "lib.rs" file row itself — a file
-        // selection (ADR 0020) groups hunks under each symbol's own
-        // signature as a section header, unlike a symbol selection (the
-        // sibling test above), which shows no header at all.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![
-                    symbol("lib.rs::foo", "foo"),
-                    ExtractedSymbol {
-                        range: LineRange { start: 10, end: 10 },
-                        ..symbol("lib.rs::bar", "bar")
-                    },
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            fan_ins: vec![],
-            file_size_warnings: vec![],
-            file_size_bands: vec![],
-            removed: vec![],
-        };
-        let app = App::new(&report);
-        let diff_text = "\
-diff --git a/lib.rs b/lib.rs
-index e69de29..4b825dc 100644
---- a/lib.rs
-+++ b/lib.rs
-@@ -1,1 +1,1 @@
--fn a() {}
-+fn foo() {}
-@@ -9,1 +10,1 @@
--fn old_bar() {}
-+fn bar() {}
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        let text = buffer_text(&terminal);
-        assert!(text.contains("fn foo()"));
-        assert!(text.contains("fn bar()"));
-        assert!(text.contains("+fn foo() {}"));
-        assert!(text.contains("+fn bar() {}"));
-    }
-
-    #[test]
-    fn should_draw_contract_header_before_hunks_when_symbol_signature_changed() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    classification: Some(Classification::SignatureChanged),
-                    previous_signature: Some("fn foo(a: i32)".to_string()),
-                    signature: "fn foo(a: i32, b: i32)".to_string(),
-                    ..symbol("lib.rs::foo", "foo")
-                }],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            fan_ins: vec![],
-            file_size_warnings: vec![],
-            file_size_bands: vec![],
-            removed: vec![],
-        };
-        // Row 0 is the "lib.rs" file row, row 1 is the "foo" symbol.
-        let app = App::new(&report).handle_key(crate::app::InputKey::Down);
-        let diff_text = "\
-diff --git a/lib.rs b/lib.rs
-index e69de29..4b825dc 100644
---- a/lib.rs
-+++ b/lib.rs
-@@ -1,1 +1,1 @@
--fn foo(a: i32) {}
-+fn foo(a: i32, b: i32) {}
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        let text = buffer_text(&terminal);
-        // The 2-line old/new contract header precedes the hunk body itself
-        // (ADR 0020's outline-before-implementation disclosure order).
-        assert!(text.contains("- fn foo(a: i32)"));
-        assert!(text.contains("+ fn foo(a: i32, b: i32)"));
-        assert!(text.contains("-fn foo(a: i32) {}"));
-        assert!(text.contains("+fn foo(a: i32, b: i32) {}"));
-    }
-
-    #[test]
-    fn should_apply_added_background_tint_and_keyword_foreground_in_diff_pane() {
-        let report = report_with_one_symbol();
-        // ADR 0020 defaults the right pane to Diff already, so no
-        // `ToggleDiff` press is needed to reach it here.
-        let app = App::new(&report).handle_key(crate::app::InputKey::Down);
-        let diff_text = "\
-diff --git a/lib.rs b/lib.rs
-index e69de29..4b825dc 100644
---- a/lib.rs
-+++ b/lib.rs
-@@ -1,1 +1,2 @@
- fn a() {}
-+fn foo() {}
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        // The added line's "fn" keyword: foreground colored by the keyword
-        // palette entry, background tinted with `ADDED_BG` — both signals
-        // present on the same cell, per ADR 0018's "fg is token color, bg is
-        // diff signal" decision. Disambiguated against the row via
-        // "+fn foo() {}" (the marker plus full added line): the left-hand
-        // tree pane's cursor row also happens to render a truncated "fn
-        // foo" label for this fixture's one symbol.
-        let keyword_style = find_cell_style(&terminal, "+fn foo() {}", "fn");
-        assert_eq!(Some(ADDED_BG), keyword_style.bg);
-        assert_eq!(Some(Color::Magenta), keyword_style.fg);
-    }
-
-    #[test]
-    fn should_apply_removed_background_tint_in_diff_pane() {
-        let report = report_with_one_symbol();
-        // ADR 0020 defaults the right pane to Diff already, so no
-        // `ToggleDiff` press is needed to reach it here.
-        let app = App::new(&report).handle_key(crate::app::InputKey::Down);
-        let diff_text = "\
-diff --git a/lib.rs b/lib.rs
-index e69de29..4b825dc 100644
---- a/lib.rs
-+++ b/lib.rs
-@@ -1,2 +1,1 @@
- fn a() {}
--fn foo() {}
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        let keyword_style = find_cell_style(&terminal, "-fn foo() {}", "fn");
-        assert_eq!(Some(REMOVED_BG), keyword_style.bg);
-        assert_eq!(Some(Color::Magenta), keyword_style.fg);
-    }
-
-    #[test]
-    fn should_keep_context_line_unstyled_background_in_diff_pane() {
-        let report = report_with_one_symbol();
-        // ADR 0020 defaults the right pane to Diff already, so no
-        // `ToggleDiff` press is needed to reach it here.
-        let app = App::new(&report).handle_key(crate::app::InputKey::Down);
-        let diff_text = "\
-diff --git a/lib.rs b/lib.rs
-index e69de29..4b825dc 100644
---- a/lib.rs
-+++ b/lib.rs
-@@ -1,1 +1,2 @@
- fn a() {}
-+fn foo() {}
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        // Context line "fn a() {}" keeps its keyword token color but must
-        // not carry either diff background tint (`Style::bg` reports an
-        // unset background as `Some(Color::Reset)`, not `None` — ratatui's
-        // own `Cell` defaults every cell's `bg` field to `Color::Reset`
-        // rather than leaving it absent). Disambiguated the same way as
-        // the added-line test above (a leading space marker rather than
-        // `+`/`-`, matching `diff_line`'s context-line rendering).
-        let context_style = find_cell_style(&terminal, " fn a() {}", "fn");
-        assert_eq!(Some(Color::Reset), context_style.bg);
-        assert_eq!(Some(Color::Magenta), context_style.fg);
-    }
-
-    #[test]
-    fn should_keep_hunk_header_dark_gray_when_diff_pane_is_highlighted() {
-        let report = report_with_one_symbol();
-        // ADR 0020 defaults the right pane to Diff already, so no
-        // `ToggleDiff` press is needed to reach it here.
-        let app = App::new(&report).handle_key(crate::app::InputKey::Down);
-        let diff_text = "\
-diff --git a/lib.rs b/lib.rs
-index e69de29..4b825dc 100644
---- a/lib.rs
-+++ b/lib.rs
-@@ -1,1 +1,2 @@
- fn a() {}
-+fn foo() {}
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        let header_style = find_cell_style(&terminal, "@@ -1,1 +1,2 @@", "@@");
-        assert_eq!(Some(Color::DarkGray), header_style.fg);
-        // DarkGray alone gives sufficient contrast; stacking `Modifier::DIM`
-        // on top of it double-dims the header to near-invisibility on many
-        // terminal themes (especially light backgrounds).
-        assert!(!header_style.add_modifier.contains(Modifier::DIM));
-    }
-
-    #[test]
-    fn should_fall_back_to_plain_diff_style_when_file_extension_is_unrecognized() {
-        // A symbol whose path has no known extension (mirrors an unbuilt
-        // language, e.g. YAML): `App::selected_diff_target` reads the path
-        // straight off the symbol/file row, so this only needs a report
-        // whose file path is unrecognized by `highlight::config_for_path`,
-        // not a real diff for an actual YAML grammar.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "config.yaml".to_string(),
-                symbols: vec![symbol("config.yaml::foo", "foo")],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            fan_ins: vec![],
-            file_size_warnings: vec![],
-            file_size_bands: vec![],
-            removed: vec![],
-        };
-        // ADR 0020 defaults the right pane to Diff already, so no
-        // `ToggleDiff` press is needed to reach it here.
-        let app = App::new(&report).handle_key(crate::app::InputKey::Down);
-        let diff_text = "\
-diff --git a/config.yaml b/config.yaml
-index e69de29..4b825dc 100644
---- a/config.yaml
-+++ b/config.yaml
-@@ -1,1 +1,2 @@
- a: 1
-+b: 2
-";
-        let diff_files = crate::diff_view::parse_diff_hunks(diff_text);
-        let diff_highlights = crate::highlight::highlight_diff_files(&diff_files);
-        let diff_content = diff_content_for(&report, &diff_files, &app);
-        let mut terminal = Terminal::new(TestBackend::new(80, 20)).expect("terminal");
-
-        terminal
-            .draw(|frame| {
-                draw(
-                    frame,
-                    &app,
-                    &report,
-                    &diff_content,
-                    &diff_highlights,
-                    &BlastRadiusSelection::NotApplicable,
-                    None,
-                );
-            })
-            .expect("draw");
-
-        let text = buffer_text(&terminal);
-        assert!(text.contains("+b: 2"));
-
-        // Falls back to the pane's original plain green foreground with no
-        // background tint at all (`Some(Color::Reset)` is ratatui's
-        // "unset" — see the context-line test above for why this isn't
-        // `None`) — highlighting failing (or, here, never applying) must
-        // never break the pre-existing diff styling.
-        let added_style = find_cell_style(&terminal, "+b: 2", "b");
-        assert_eq!(Some(Color::Reset), added_style.bg);
-        assert_eq!(Some(Color::Green), added_style.fg);
-    }
-}
+#[path = "diff_pane_tests/mod.rs"]
+mod tests;
